added better json-ld
