+++ components/layout/layout.tsx
@@ -10,24 +10,13 @@ import dayjs from "dayjs";
 import dynamic from "next/dynamic";
 import { Open_Sans } from "next/font/google";
 import { useReportWebVitals } from "next/web-vitals";
-import { WebSite, WithContext } from "schema-dts";
 import { MegaMenuLayout, NavMenuGroup } from "ssw.megamenu";
-import layoutData from "../../content/global/index.json";
 import { CustomLink } from "../customLink";
 
 export const openSans = Open_Sans({
   variable: "--open-sans-font",
   subsets: ["latin"],
 });
-
-const structuredData: WithContext<WebSite> = {
-  "@context": "https://schema.org",
-  "@type": "WebSite",
-  name: layoutData.header.site_name,
-  alternateName: layoutData.header.alternate_site_name,
-  description: layoutData.header.description,
-  url: layoutData.header.url,
-};
 
 const LiveStreamWidget = dynamic(
   () => {
@@ -132,13 +121,6 @@ export const Layout = ({
         <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#cc4141" />
         <meta name="msapplication-TileColor" content="#cc4141" />
         <meta name="theme-color" content="#ffffff" />
-
-        {router.query.filename === "home" && (
-          <script
-            type="application/ld+json"
-            dangerouslySetInnerHTML={{ __html: JSON.stringify(structuredData) }}
-          />
-        )}
       </Head>
       <Theme>
         {/* Ensures next/font CSS variable is accessible for all components */}
